fix(data-warehouse): update GitHub tests for resumable source refactor
Rewrites `test_github_source.py` to exercise the new hand-written
paginated fetch loop: removes tests for the deleted `GithubPaginator`
class and `get_resource` builder, adds coverage for `_build_initial_params`,
`_build_initial_url`, `_parse_next_url`, `_is_older_than_cutoff`,
`_should_stop_desc`, and `get_rows` fresh-run and resume paths.

Removes the duplicate per-source `test_github.py` in favour of the
canonical tests location under `posthog/temporal/tests/data_imports/github/`.

Fixes mypy import errors for removed symbols blocking Python code quality
and Django Temporal CI checks.

Generated-By: PostHog Code
Task-Id: f0c1e556-f8f6-4e20-bd6d-6e88bf2cd59e
