+++ posthog/temporal/tests/data_imports/github/test_github_source.py
@@ -6,33 +6,57 @@ from unittest import mock
 import requests
 from parameterized import parameterized
 
-from posthog.temporal.data_imports.sources.common.rest_source.typing import EndpointResource
 from posthog.temporal.data_imports.sources.github.github import (
-    GithubPaginator,
+    GithubResumeConfig,
+    _build_initial_params,
+    _build_initial_url,
     _flatten_commit,
     _flatten_stargazer,
     _format_incremental_value,
     _is_issue_not_pr,
-    get_resource,
+    _is_older_than_cutoff,
+    _parse_next_url,
+    _should_stop_desc,
+    get_rows,
     github_source,
     validate_credentials,
 )
+from posthog.temporal.data_imports.sources.github.settings import GITHUB_ENDPOINTS
 
 
-def _endpoint_params(resource: EndpointResource) -> dict[str, Any]:
-    endpoint = resource.get("endpoint")
-    assert isinstance(endpoint, dict)
-    params = endpoint.get("params")
-    assert isinstance(params, dict)
-    return params
+def _make_response(status: int = 200, body: Any = None, link: str = "") -> mock.Mock:
+    resp = mock.Mock()
+    resp.status_code = status
+    resp.ok = 200 <= status < 300
+    resp.json.return_value = body if body is not None else []
+    resp.headers = {"Link": link} if link else {}
+    return resp
 
 
-def _endpoint_path(resource: EndpointResource) -> str:
-    endpoint = resource.get("endpoint")
-    assert isinstance(endpoint, dict)
-    path = endpoint.get("path")
-    assert isinstance(path, str)
-    return path
+def _make_manager(*, can_resume: bool = False, resume_state: GithubResumeConfig | None = None) -> mock.Mock:
+    manager = mock.Mock()
+    manager.can_resume.return_value = can_resume
+    manager.load_state.return_value = resume_state
+    manager.save_state = mock.Mock()
+    return manager
+
+
+class _ImmediateBatcher:
+    """Test double for Batcher that emits every batched item as its own chunk."""
+
+    def __init__(self) -> None:
+        self._item: Any = None
+
+    def batch(self, item: Any) -> None:
+        self._item = item
+
+    def should_yield(self, include_incomplete_chunk: bool = False) -> bool:
+        return self._item is not None
+
+    def get_table(self) -> Any:
+        item = self._item
+        self._item = None
+        return item
 
 
 class TestFormatIncrementalValue:
@@ -44,49 +68,50 @@ class TestFormatIncrementalValue:
             ("integer_passthrough", 1737100800, "1737100800"),
         ]
     )
-    def test_formats_value(self, _name, value, expected):
+    def test_formats_value(self, _name: str, value: Any, expected: str) -> None:
         assert _format_incremental_value(value) == expected
 
 
-class TestGetResource:
-    def test_issues_full_refresh(self):
-        resource = get_resource("issues", "owner/repo", should_use_incremental_field=False)
-        params = _endpoint_params(resource)
+class TestBuildInitialParams:
+    def test_issues_full_refresh_defaults(self) -> None:
+        params = _build_initial_params(
+            GITHUB_ENDPOINTS["issues"],
+            "issues",
+            should_use_incremental_field=False,
+            db_incremental_field_last_value=None,
+            incremental_field=None,
+        )
 
-        assert resource["name"] == "issues"
-        assert resource["primary_key"] == "id"
-        assert resource["write_disposition"] == "replace"
-        assert _endpoint_path(resource) == "/repos/owner/repo/issues"
+        assert params["sort"] == "created"
+        assert params["direction"] == "asc"
         assert params["state"] == "all"
         assert params["per_page"] == 100
         assert "since" not in params
 
-    def test_issues_incremental_with_since(self):
+    def test_issues_incremental_with_since(self) -> None:
         last_value = datetime(2026, 1, 15, 10, 0, 0, tzinfo=UTC)
-        resource = get_resource(
+        params = _build_initial_params(
+            GITHUB_ENDPOINTS["issues"],
             "issues",
-            "owner/repo",
             should_use_incremental_field=True,
             db_incremental_field_last_value=last_value,
+            incremental_field="updated_at",
         )
-        params = _endpoint_params(resource)
 
-        assert resource["write_disposition"] == {"disposition": "merge", "strategy": "upsert"}
         assert params["since"] == "2026-01-15T10:00:00+00:00"
         assert params["sort"] == "updated"
         assert params["direction"] == "asc"
 
-    def test_commits_incremental_with_since(self):
+    def test_commits_incremental_desc(self) -> None:
         last_value = datetime(2026, 1, 15, 10, 0, 0, tzinfo=UTC)
-        resource = get_resource(
+        params = _build_initial_params(
+            GITHUB_ENDPOINTS["commits"],
             "commits",
-            "owner/repo",
             should_use_incremental_field=True,
             db_incremental_field_last_value=last_value,
+            incremental_field="created_at",
         )
-        params = _endpoint_params(resource)
 
-        assert resource["primary_key"] == "sha"
         assert params["since"] == "2026-01-15T10:00:00+00:00"
         assert params["sort"] == "created"
         assert params["direction"] == "desc"
@@ -98,154 +123,65 @@ class TestGetResource:
             ("default_to_updated", None, "updated"),
         ]
     )
-    def test_pull_requests_incremental_sort_with_cutoff(self, _name, incremental_field, expected_sort):
+    def test_pull_requests_incremental_sort_with_cutoff(
+        self, _name: str, incremental_field: str | None, expected_sort: str
+    ) -> None:
         last_value = datetime(2026, 1, 15, 10, 0, 0, tzinfo=UTC)
-        resource = get_resource(
+        params = _build_initial_params(
+            GITHUB_ENDPOINTS["pull_requests"],
             "pull_requests",
-            "owner/repo",
             should_use_incremental_field=True,
-            incremental_field=incremental_field,
             db_incremental_field_last_value=last_value,
+            incremental_field=incremental_field,
         )
-        params = _endpoint_params(resource)
 
         assert params["sort"] == expected_sort
         assert params["direction"] == "desc"
-        assert "since" not in params
+        assert "since" not in params  # pull_requests does not support 'since'
 
-    def test_pull_requests_incremental_created_asc_without_cutoff(self):
-        resource = get_resource(
-            "pull_requests",
-            "owner/repo",
-            should_use_incremental_field=True,
-            db_incremental_field_last_value=None,
-        )
-        params = _endpoint_params(resource)
-
-        assert params["sort"] == "created"
-        assert params["direction"] == "asc"
-
-    def test_pull_requests_full_refresh_created_asc(self):
-        resource = get_resource("pull_requests", "owner/repo", should_use_incremental_field=False)
-        params = _endpoint_params(resource)
-
-        assert params["sort"] == "created"
-        assert params["direction"] == "asc"
-
-    def test_issues_incremental_no_since_without_last_value(self):
-        resource = get_resource(
+    def test_issues_incremental_no_since_without_last_value(self) -> None:
+        params = _build_initial_params(
+            GITHUB_ENDPOINTS["issues"],
             "issues",
-            "owner/repo",
             should_use_incremental_field=True,
             db_incremental_field_last_value=None,
+            incremental_field=None,
         )
-        params = _endpoint_params(resource)
 
-        assert resource["write_disposition"] == {"disposition": "merge", "strategy": "upsert"}
         assert "since" not in params
         assert params["sort"] == "created"
         assert params["direction"] == "asc"
 
 
-class TestGithubPaginator:
-    def _make_response(self, link_header: str = "") -> mock.MagicMock:
-        response = mock.MagicMock()
-        response.headers = {"Link": link_header} if link_header else {}
-        return response
+class TestBuildInitialUrl:
+    def test_with_params(self) -> None:
+        url = _build_initial_url(GITHUB_ENDPOINTS["issues"], "owner/repo", {"per_page": 100, "state": "all"})
+        assert url == "https://api.github.com/repos/owner/repo/issues?per_page=100&state=all"
 
-    def test_no_link_header(self):
-        paginator = GithubPaginator()
-        paginator.update_state(self._make_response(), data=[{"id": 1}])
+    def test_no_params(self) -> None:
+        url = _build_initial_url(GITHUB_ENDPOINTS["issues"], "owner/repo", {})
+        assert url == "https://api.github.com/repos/owner/repo/issues"
 
-        assert paginator.has_next_page is False
 
-    def test_next_page_from_link_header(self):
-        link = '<https://api.github.com/repos/owner/repo/issues?page=2>; rel="next", <https://api.github.com/repos/owner/repo/issues?page=5>; rel="last"'
-        paginator = GithubPaginator()
-        paginator.update_state(self._make_response(link), data=[{"id": 1}])
-
-        assert paginator.has_next_page is True
-        assert paginator._next_url == "https://api.github.com/repos/owner/repo/issues?page=2"
-
-    def test_no_next_rel_in_link_header(self):
-        link = '<https://api.github.com/repos/owner/repo/issues?page=1>; rel="prev"'
-        paginator = GithubPaginator()
-        paginator.update_state(self._make_response(link), data=[{"id": 1}])
-
-        assert paginator.has_next_page is False
-
-    def test_desc_sort_stops_on_old_records(self):
-        cutoff = datetime(2026, 1, 20, 0, 0, 0, tzinfo=UTC)
-        link = '<https://api.github.com/next>; rel="next"'
-        paginator = GithubPaginator(
-            incremental_field="updated_at",
-            db_incremental_field_last_value=cutoff,
-            sort_mode="desc",
-        )
-
-        data = [
-            {"updated_at": "2026-01-25T10:00:00Z"},
-            {"updated_at": "2026-01-19T10:00:00Z"},  # older than cutoff
+class TestParseNextUrl:
+    @parameterized.expand(
+        [
+            ("empty", "", None),
+            (
+                "next_and_last",
+                '<https://api.github.com/x?page=2>; rel="next", <https://api.github.com/x?page=5>; rel="last"',
+                "https://api.github.com/x?page=2",
+            ),
+            ("only_last", '<https://api.github.com/x?page=5>; rel="last"', None),
+            (
+                "prev_and_next",
+                '<https://api.github.com/x?page=1>; rel="prev", <https://api.github.com/x?page=3>; rel="next"',
+                "https://api.github.com/x?page=3",
+            ),
         ]
-        paginator.update_state(self._make_response(link), data=data)
-
-        assert paginator.has_next_page is False
-
-    def test_desc_sort_continues_when_all_newer(self):
-        cutoff = datetime(2026, 1, 10, 0, 0, 0, tzinfo=UTC)
-        link = '<https://api.github.com/next>; rel="next"'
-        paginator = GithubPaginator(
-            incremental_field="updated_at",
-            db_incremental_field_last_value=cutoff,
-            sort_mode="desc",
-        )
-
-        data = [
-            {"updated_at": "2026-01-25T10:00:00Z"},
-            {"updated_at": "2026-01-15T10:00:00Z"},
-        ]
-        paginator.update_state(self._make_response(link), data=data)
-
-        assert paginator.has_next_page is True
-
-    def test_asc_sort_ignores_cutoff(self):
-        cutoff = datetime(2026, 1, 20, 0, 0, 0, tzinfo=UTC)
-        link = '<https://api.github.com/next>; rel="next"'
-        paginator = GithubPaginator(
-            incremental_field="updated_at",
-            db_incremental_field_last_value=cutoff,
-            sort_mode="asc",
-        )
-
-        data = [{"updated_at": "2026-01-10T10:00:00Z"}]
-        paginator.update_state(self._make_response(link), data=data)
-
-        assert paginator.has_next_page is True
-
-    def test_desc_sort_no_cutoff_continues(self):
-        link = '<https://api.github.com/next>; rel="next"'
-        paginator = GithubPaginator(
-            incremental_field="updated_at",
-            db_incremental_field_last_value=None,
-            sort_mode="desc",
-        )
-
-        data = [{"updated_at": "2026-01-10T10:00:00Z"}]
-        paginator.update_state(self._make_response(link), data=data)
-
-        assert paginator.has_next_page is True
-
-    def test_update_request_sets_next_url(self):
-        link = '<https://api.github.com/repos/owner/repo/issues?page=2>; rel="next"'
-        paginator = GithubPaginator()
-        paginator.update_state(self._make_response(link), data=[{"id": 1}])
-
-        request = mock.MagicMock()
-        request.params = {"per_page": 100}
-        paginator.update_request(request)
-
-        assert request.url == "https://api.github.com/repos/owner/repo/issues?page=2"
-        assert request.params == {}
+    )
+    def test_parse(self, _name: str, header: str, expected: str | None) -> None:
+        assert _parse_next_url(header) == expected
 
 
 class TestIsOlderThanCutoff:
@@ -253,50 +189,83 @@ class TestIsOlderThanCutoff:
         [
             ("z_suffix_older", "2026-01-15T10:00:00Z", True),
             ("offset_older", "2026-01-15T10:00:00+00:00", True),
-            ("equal_to_cutoff", "2026-01-20T00:00:00Z", True),
+            ("equal_to_cutoff", "2026-01-20T00:00:00+00:00", True),
             ("newer_than_cutoff", "2026-01-25T10:00:00Z", False),
             ("none_value", None, False),
             ("invalid_string", "not-a-date", False),
         ]
     )
-    def test_string_comparison(self, _name, value, expected):
+    def test_string_comparison(self, _name: str, value: Any, expected: bool) -> None:
         cutoff = datetime(2026, 1, 20, 0, 0, 0, tzinfo=UTC)
-        paginator = GithubPaginator(db_incremental_field_last_value=cutoff)
-        assert paginator._is_older_than_cutoff(value) == expected
+        assert _is_older_than_cutoff(value, cutoff) is expected
 
-    def test_datetime_comparison(self):
+    def test_datetime_comparison(self) -> None:
         cutoff = datetime(2026, 1, 20, 0, 0, 0, tzinfo=UTC)
-        paginator = GithubPaginator(db_incremental_field_last_value=cutoff)
+        assert _is_older_than_cutoff(datetime(2026, 1, 15, 0, 0, 0, tzinfo=UTC), cutoff) is True
+        assert _is_older_than_cutoff(datetime(2026, 1, 25, 0, 0, 0, tzinfo=UTC), cutoff) is False
 
-        assert paginator._is_older_than_cutoff(datetime(2026, 1, 15, 0, 0, 0, tzinfo=UTC)) is True
-        assert paginator._is_older_than_cutoff(datetime(2026, 1, 25, 0, 0, 0, tzinfo=UTC)) is False
 
-    def test_none_cutoff(self):
-        paginator = GithubPaginator(db_incremental_field_last_value=None)
-        assert paginator._is_older_than_cutoff("2026-01-15T10:00:00Z") is False
+class TestShouldStopDesc:
+    def test_asc_mode_never_stops(self) -> None:
+        assert (
+            _should_stop_desc(
+                data=[{"updated_at": "2000-01-01T00:00:00Z"}],
+                sort_mode="asc",
+                incremental_field="updated_at",
+                cutoff=datetime(2026, 3, 4, tzinfo=UTC),
+            )
+            is False
+        )
+
+    def test_desc_stops_on_old_record(self) -> None:
+        assert (
+            _should_stop_desc(
+                data=[
+                    {"updated_at": "2026-05-01T00:00:00Z"},
+                    {"updated_at": "2025-01-01T00:00:00Z"},
+                ],
+                sort_mode="desc",
+                incremental_field="updated_at",
+                cutoff=datetime(2026, 3, 4, tzinfo=UTC),
+            )
+            is True
+        )
+
+    def test_desc_continues_when_all_newer(self) -> None:
+        assert (
+            _should_stop_desc(
+                data=[{"updated_at": "2026-05-01T00:00:00Z"}],
+                sort_mode="desc",
+                incremental_field="updated_at",
+                cutoff=datetime(2026, 3, 4, tzinfo=UTC),
+            )
+            is False
+        )
+
+    def test_desc_no_cutoff_continues(self) -> None:
+        assert (
+            _should_stop_desc(
+                data=[{"updated_at": "2026-05-01T00:00:00Z"}],
+                sort_mode="desc",
+                incremental_field="updated_at",
+                cutoff=None,
+            )
+            is False
+        )
 
 
 class TestFlattenCommit:
-    def test_flattens_nested_commit_data(self):
+    def test_flattens_nested_commit_data(self) -> None:
         item = {
             "sha": "abc123",
             "commit": {
                 "message": "Fix bug",
-                "author": {
-                    "name": "Alice",
-                    "email": "alice@example.com",
-                    "date": "2026-01-10T10:00:00Z",
-                },
-                "committer": {
-                    "name": "Bob",
-                    "email": "bob@example.com",
-                    "date": "2026-01-10T11:00:00Z",
-                },
+                "author": {"name": "Alice", "email": "alice@example.com", "date": "2026-01-10T10:00:00Z"},
+                "committer": {"name": "Bob", "email": "bob@example.com", "date": "2026-01-10T11:00:00Z"},
             },
             "author": {"id": 100, "login": "alice"},
             "committer": {"id": 101, "login": "bob"},
         }
-
         result = _flatten_commit(item)
 
         assert result["message"] == "Fix bug"
@@ -311,17 +280,15 @@ class TestFlattenCommit:
         assert result["committer_id"] == 101
         assert result["committer_login"] == "bob"
 
-    def test_handles_missing_nested_data(self):
+    def test_handles_missing_nested_data(self) -> None:
         item = {"sha": "abc123"}
         result = _flatten_commit(item)
 
         assert result == {"sha": "abc123"}
-        assert "message" not in result
-        assert "author_name" not in result
 
 
 class TestFlattenStargazer:
-    def test_flattens_user_data(self):
+    def test_flattens_user_data(self) -> None:
         item = {
             "starred_at": "2026-01-10T10:00:00Z",
             "user": {
@@ -331,7 +298,6 @@ class TestFlattenStargazer:
                 "type": "User",
             },
         }
-
         result = _flatten_stargazer(item)
 
         assert result["id"] == 100
@@ -341,10 +307,9 @@ class TestFlattenStargazer:
         assert result["starred_at"] == "2026-01-10T10:00:00Z"
         assert "user" not in result
 
-    def test_handles_missing_user(self):
+    def test_handles_missing_user(self) -> None:
         item = {"starred_at": "2026-01-10T10:00:00Z"}
         result = _flatten_stargazer(item)
-
         assert result == {"starred_at": "2026-01-10T10:00:00Z"}
 
 
@@ -356,12 +321,12 @@ class TestIsIssueNotPr:
             ("pr_null", {"id": 3, "pull_request": None}, True),
         ]
     )
-    def test_filters_correctly(self, _name, item, expected):
+    def test_filters_correctly(self, _name: str, item: dict[str, Any], expected: bool) -> None:
         assert _is_issue_not_pr(item) == expected
 
 
 class TestValidateCredentials:
-    def test_valid_credentials(self):
+    def test_valid_credentials(self) -> None:
         with mock.patch("posthog.temporal.data_imports.sources.github.github.requests.get") as mock_get:
             mock_get.return_value = mock.MagicMock(status_code=200)
             valid, error = validate_credentials("token", "owner/repo")
@@ -375,7 +340,7 @@ class TestValidateCredentials:
             ("not_found", 404, "Repository 'owner/repo' not found or not accessible"),
         ]
     )
-    def test_error_status_codes(self, _name, status_code, expected_message):
+    def test_error_status_codes(self, _name: str, status_code: int, expected_message: str) -> None:
         with mock.patch("posthog.temporal.data_imports.sources.github.github.requests.get") as mock_get:
             mock_get.return_value = mock.MagicMock(status_code=status_code)
             valid, error = validate_credentials("token", "owner/repo")
@@ -383,7 +348,7 @@ class TestValidateCredentials:
         assert valid is False
         assert error == expected_message
 
-    def test_json_error_response(self):
+    def test_json_error_response(self) -> None:
         with mock.patch("posthog.temporal.data_imports.sources.github.github.requests.get") as mock_get:
             mock_response = mock.MagicMock(status_code=403)
             mock_response.json.return_value = {"message": "API rate limit exceeded"}
@@ -393,7 +358,7 @@ class TestValidateCredentials:
         assert valid is False
         assert error == "API rate limit exceeded"
 
-    def test_request_exception(self):
+    def test_request_exception(self) -> None:
         with mock.patch("posthog.temporal.data_imports.sources.github.github.requests.get") as mock_get:
             mock_get.side_effect = requests.exceptions.ConnectionError("Connection refused")
             valid, error = validate_credentials("token", "owner/repo")
@@ -402,7 +367,7 @@ class TestValidateCredentials:
         assert error is not None
         assert "Connection refused" in error
 
-    def test_sends_correct_headers(self):
+    def test_sends_correct_headers(self) -> None:
         with mock.patch("posthog.temporal.data_imports.sources.github.github.requests.get") as mock_get:
             mock_get.return_value = mock.MagicMock(status_code=200)
             validate_credentials("my-token", "owner/repo")
@@ -418,18 +383,21 @@ class TestValidateCredentials:
 class TestGithubSourceSortMode:
     """SourceResponse.sort_mode must match the actual API request direction."""
 
-    def _make_source(self, endpoint, should_use_incremental_field, db_incremental_field_last_value=None):
-        with mock.patch("posthog.temporal.data_imports.sources.github.github.rest_api_resource") as mock_resource_fn:
-            mock_resource_fn.return_value = mock.MagicMock()
-            return github_source(
-                personal_access_token="token",
-                repository="owner/repo",
-                endpoint=endpoint,
-                team_id=1,
-                job_id="job-1",
-                should_use_incremental_field=should_use_incremental_field,
-                db_incremental_field_last_value=db_incremental_field_last_value,
-            )
+    def _make_source(
+        self,
+        endpoint: str,
+        should_use_incremental_field: bool,
+        db_incremental_field_last_value: Any = None,
+    ) -> Any:
+        return github_source(
+            personal_access_token="token",
+            repository="owner/repo",
+            endpoint=endpoint,
+            logger=mock.Mock(),
+            resumable_source_manager=_make_manager(),
+            should_use_incremental_field=should_use_incremental_field,
+            db_incremental_field_last_value=db_incremental_field_last_value,
+        )
 
     @parameterized.expand(
         [
@@ -443,10 +411,141 @@ class TestGithubSourceSortMode:
                 "desc",
             ),
             ("commits_first_sync_no_cutoff", "commits", True, None, "asc"),
-            ("commits_incremental_with_cutoff", "commits", True, datetime(2026, 1, 15, tzinfo=UTC), "desc"),
+            (
+                "commits_incremental_with_cutoff",
+                "commits",
+                True,
+                datetime(2026, 1, 15, tzinfo=UTC),
+                "desc",
+            ),
             ("issues_always_asc", "issues", True, datetime(2026, 1, 15, tzinfo=UTC), "asc"),
         ]
     )
-    def test_sort_mode(self, _name, endpoint, incremental, cutoff, expected_sort_mode):
+    def test_sort_mode(
+        self,
+        _name: str,
+        endpoint: str,
+        incremental: bool,
+        cutoff: Any,
+        expected_sort_mode: str,
+    ) -> None:
         response = self._make_source(endpoint, incremental, cutoff)
         assert response.sort_mode == expected_sort_mode
+
+
+class TestGetRowsResume:
+    """Integration tests for the fresh-run and resume flows in get_rows."""
+
+    def _patch_batcher(self) -> Any:
+        return mock.patch(
+            "posthog.temporal.data_imports.sources.github.github.Batcher",
+            autospec=False,
+            side_effect=lambda logger, chunk_size, chunk_size_bytes: _ImmediateBatcher(),
+        )
+
+    def test_fresh_run_saves_next_url_after_batch(self) -> None:
+        manager = _make_manager(can_resume=False)
+        page1 = [{"id": 1, "updated_at": "2026-01-01T00:00:00Z"}]
+        page2 = [{"id": 2, "updated_at": "2026-01-02T00:00:00Z"}]
+        link_page1 = '<https://api.github.com/repos/owner/repo/releases?page=2>; rel="next"'
+        responses = [
+            _make_response(body=page1, link=link_page1),
+            _make_response(body=page2, link=""),
+        ]
+
+        with (
+            self._patch_batcher(),
+            mock.patch(
+                "posthog.temporal.data_imports.sources.github.github.requests.get",
+                side_effect=responses,
+            ) as mock_get,
+        ):
+            list(
+                get_rows(
+                    personal_access_token="tok",
+                    repository="owner/repo",
+                    endpoint="releases",
+                    logger=mock.Mock(),
+                    resumable_source_manager=manager,
+                    should_use_incremental_field=False,
+                )
+            )
+
+        first_url = mock_get.call_args_list[0].args[0]
+        assert first_url.startswith("https://api.github.com/repos/owner/repo/releases")
+        manager.load_state.assert_not_called()
+        assert manager.save_state.call_count == 1
+        saved = manager.save_state.call_args.args[0]
+        assert isinstance(saved, GithubResumeConfig)
+        assert saved.next_url == "https://api.github.com/repos/owner/repo/releases?page=2"
+
+    def test_resume_uses_saved_url(self) -> None:
+        saved_url = "https://api.github.com/repos/owner/repo/releases?page=4"
+        manager = _make_manager(can_resume=True, resume_state=GithubResumeConfig(next_url=saved_url))
+        responses = [_make_response(body=[{"id": 42, "updated_at": "2026-01-01T00:00:00Z"}], link="")]
+
+        with (
+            self._patch_batcher(),
+            mock.patch(
+                "posthog.temporal.data_imports.sources.github.github.requests.get",
+                side_effect=responses,
+            ) as mock_get,
+        ):
+            list(
+                get_rows(
+                    personal_access_token="tok",
+                    repository="owner/repo",
+                    endpoint="releases",
+                    logger=mock.Mock(),
+                    resumable_source_manager=manager,
+                    should_use_incremental_field=False,
+                )
+            )
+
+        assert mock_get.call_args_list[0].args[0] == saved_url
+        manager.load_state.assert_called_once()
+
+    def test_empty_first_page_ends_loop(self) -> None:
+        manager = _make_manager(can_resume=False)
+        with (
+            self._patch_batcher(),
+            mock.patch(
+                "posthog.temporal.data_imports.sources.github.github.requests.get",
+                side_effect=[_make_response(body=[], link="")],
+            ),
+        ):
+            rows = list(
+                get_rows(
+                    personal_access_token="tok",
+                    repository="owner/repo",
+                    endpoint="releases",
+                    logger=mock.Mock(),
+                    resumable_source_manager=manager,
+                    should_use_incremental_field=False,
+                )
+            )
+
+        assert rows == []
+        manager.save_state.assert_not_called()
+
+    def test_final_page_does_not_save_state(self) -> None:
+        manager = _make_manager(can_resume=False)
+        with (
+            self._patch_batcher(),
+            mock.patch(
+                "posthog.temporal.data_imports.sources.github.github.requests.get",
+                side_effect=[_make_response(body=[{"id": 1}], link="")],
+            ),
+        ):
+            list(
+                get_rows(
+                    personal_access_token="tok",
+                    repository="owner/repo",
+                    endpoint="releases",
+                    logger=mock.Mock(),
+                    resumable_source_manager=manager,
+                    should_use_incremental_field=False,
+                )
+            )
+
+        manager.save_state.assert_not_called()
